[Core] DisposeBase: Set IsDisposed to true after calling Destroy. [PDX-2457]
Note: If it is set before, it is always true at the beginning of the override of Destroy and cannot be used to know if the object has already been disposed. Then we need to create another local variable just to know the dispose status inside the Destroy function which is really inconvenient.