[General] Update Microsoft.CodeAnalysis to version 1.1
While building with VS 2015 update 1, we were getting the following
warning:

CSC : warning AD0001: Analyzer 'Microsoft.CodeAnalysis.CSharp.Analyzers.FixAnalyzers.CSharpFixerWithFixAllAnalyzer' threw an exception of type 'System.TypeInitializationException' with message 'The type initializer for 'Microsoft.CodeAnalysis.Analyzers.FixAnalyzers.FixerWithFixAllAnalyzer`1' threw an exception.'.

The solution we found online was to remove that warning was to update
the Microsoft.CodeAnalysis to version 1.1. This also required updating
nuget to version 3.0 or above (we chose the latest 3.3). Updated various
.config and .csproj to refer to the new assembly versions.