[Graphics] Remove explicit reference to System.Drawing/System.Windows.Forms
Since we do not use those assemblies directly, we do not need to include
them. However due to a change in C# in VS 2015 Updated 2 (see Github
issue https://github.com/dotnet/roslyn/issues/9370), we are now using
the GL.BlendFactor that has no overload to avoid the CS0012 compilation
error for not including System.Drawing which we do not use directly.