[PackageManager] Move NugetStore and dependents to a new project
Updated client projects to refer to the new PackageManager project. We
chose this path instead of a shared project because when we will upgrade
to Nuget v3.0 each dependent project would also need to add 10+ Nuget
assemblies. Also in the future, we won't expose the fact we are actually
using Nuget for our package manager.