[Launcher] Fix self-update that would not copy the files from the package
This was due because `GetFiles` actually extract the files in the temp
folder and then `GetStream` should open the stream from the temp folder
and not from the installed path.

We might want to review this in the future so that it gets installed as
any other package and then copied to where we want.